refactor: extract agent turn runner
